requirements.txt
Django==1.7.1
djangorestframework==2.4.4
oauthlib==0.7.1
psycopg2==2.5.4
python-linkedin==4.1
requests==2.4.3
requests-oauthlib==0.4.2
wsgiref==0.1.2
